Moved LibraryCustomerBuilder back to customer package
