+++ src/main/java/library/customer/LibraryCustomerBuilder.java
@@ -1,55 +1,53 @@
-package library.builder;
-
-import library.customer.LibraryCustomer;
+package library.customer;
 
 import java.time.LocalDate;
 
-public class LibraryCustomerBuilder {
+class LibraryCustomerBuilder {
 
     private String name;
     private String email;
     private LocalDate dateOfBirth = LocalDate.now();
     private String[] address = new String[4];
 
-    public LibraryCustomerBuilder() {
+    LibraryCustomerBuilder() {
         //Builder class, empty constructor is required
     }
 
-    public LibraryCustomer buildCustomer() {
+    LibraryCustomer buildCustomer() {
         return new LibraryCustomer(name, email, dateOfBirth, address[0], address[1], address[2], address[3]);
     }
 
-    public LibraryCustomerBuilder name(String name) {
+    LibraryCustomerBuilder name(String name) {
         this.name = name;
         return this;
     }
 
-    public LibraryCustomerBuilder email(String email) {
+    LibraryCustomerBuilder email(String email) {
         this.email = email;
         return this;
     }
 
-    public LibraryCustomerBuilder dob(LocalDate dateOfBirth) {
+    LibraryCustomerBuilder dob(LocalDate dateOfBirth) {
         this.dateOfBirth = dateOfBirth;
         return this;
     }
 
-    public LibraryCustomerBuilder address1(String address1) {
+    LibraryCustomerBuilder address1(String address1) {
         this.address[0] = address1;
         return this;
     }
 
-    public LibraryCustomerBuilder address2(String address2) {
+    LibraryCustomerBuilder address2(String address2) {
         this.address[1] = address2;
         return this;
     }
 
-    public LibraryCustomerBuilder address3(String address3) {
+    LibraryCustomerBuilder address3(String address3) {
         this.address[2] = address3;
         return this;
     }
 
-    public LibraryCustomerBuilder postCode(String postCode) {
+    LibraryCustomerBuilder postCode(String postCode) {
         this.address[3] = postCode;
         return this;
     }
